Add files via upload
First version of the site map for the Term Project website. Currently, it is a very general overview of what the site structure will be like. Also note that the actual site currently doesn't correspond to the site map as only the home page and the CSS stylesheet have been added so far, but the site will be updated soon to resolve this.